@@ -4,6 +4,7 @@ require 'savon'
 require 'netsuite/version'
 require 'netsuite/errors'
 require 'netsuite/utilities'
+require 'netsuite/utilities/data_center'
 require 'netsuite/rest/utilities/roles'
 require 'netsuite/rest/utilities/request'
 require 'netsuite/core_ext/string/lower_camelcase'
@@ -51,11 +52,13 @@ module NetSuite
     autoload :Delete,           'netsuite/actions/delete'
     autoload :DeleteList,       'netsuite/actions/delete_list'
     autoload :Get,              'netsuite/actions/get'
+    autoload :GetDeleted,       'netsuite/actions/get_deleted'
     autoload :GetAll,           'netsuite/actions/get_all'
     autoload :GetList,          'netsuite/actions/get_list'
     autoload :GetSelectValue,   'netsuite/actions/get_select_value'
     autoload :Initialize,       'netsuite/actions/initialize'
     autoload :Update,           'netsuite/actions/update'
+    autoload :UpdateList,       'netsuite/actions/update_list'
     autoload :Upsert,           'netsuite/actions/upsert'
     autoload :UpsertList,       'netsuite/actions/upsert_list'
     autoload :Search,           'netsuite/actions/search'
@@ -64,6 +67,10 @@ module NetSuite
 
   module Records
     autoload :AssemblyItem,                     'netsuite/records/assembly_item'
+    autoload :AssemblyBuild,                    'netsuite/records/assembly_build'
+    autoload :AssemblyComponent,                'netsuite/records/assembly_component'
+    autoload :AssemblyComponentList,            'netsuite/records/assembly_component_list'
+    autoload :AssemblyUnbuild,                  'netsuite/records/assembly_unbuild'
     autoload :Account,                          'netsuite/records/account'
     autoload :AccountingPeriod,                 'netsuite/records/accounting_period'
     autoload :Address,                          'netsuite/records/address'
@@ -74,7 +81,12 @@ module NetSuite
     autoload :BillingScheduleMilestoneList,     'netsuite/records/billing_schedule_milestone_list'
     autoload :BillingScheduleRecurrence,        'netsuite/records/billing_schedule_recurrence'
     autoload :BillingScheduleRecurrenceList,    'netsuite/records/billing_schedule_recurrence_list'
+    autoload :Bin,                              'netsuite/records/bin'
+    autoload :BinNumber,                        'netsuite/records/bin_number'
     autoload :BinNumberList,                    'netsuite/records/bin_number_list'
+    autoload :BinTransfer,                      'netsuite/records/bin_transfer'
+    autoload :BinTransferInventory,             'netsuite/records/bin_transfer_inventory'
+    autoload :BinTransferInventoryList,         'netsuite/records/bin_transfer_inventory_list'
     autoload :CashSale,                         'netsuite/records/cash_sale'
     autoload :CashSaleItem,                     'netsuite/records/cash_sale_item'
     autoload :CashSaleItemList,                 'netsuite/records/cash_sale_item_list'
@@ -100,21 +112,33 @@ module NetSuite
     autoload :CustomerAddressbook,              'netsuite/records/customer_addressbook'
     autoload :CustomerAddressbookList,          'netsuite/records/customer_addressbook_list'
     autoload :CustomerCategory,                 'netsuite/records/customer_category'
+    autoload :CustomerCreditCards,              'netsuite/records/customer_credit_cards'
+    autoload :CustomerCreditCardsList,          'netsuite/records/customer_credit_cards_list'
     autoload :CustomerCurrency,                 'netsuite/records/customer_currency'
     autoload :CustomerCurrencyList,             'netsuite/records/customer_currency_list'
     autoload :CustomerDeposit,                  'netsuite/records/customer_deposit'
+    autoload :CustomerDepositApplyList,         'netsuite/records/customer_deposit_apply_list'
+    autoload :CustomerDepositApply,             'netsuite/records/customer_deposit_apply'
     autoload :CustomerPartnersList,             'netsuite/records/customer_partners_list'
     autoload :CustomerPayment,                  'netsuite/records/customer_payment'
     autoload :CustomerPaymentApply,             'netsuite/records/customer_payment_apply'
     autoload :CustomerPaymentApplyList,         'netsuite/records/customer_payment_apply_list'
+    autoload :CustomerPartner,                  'netsuite/records/customer_partner'
     autoload :CustomerRefund,                   'netsuite/records/customer_refund'
     autoload :CustomerRefundApply,              'netsuite/records/customer_refund_apply'
     autoload :CustomerRefundApplyList,          'netsuite/records/customer_refund_apply_list'
     autoload :CustomerRefundDeposit,            'netsuite/records/customer_refund_deposit'
     autoload :CustomerRefundDepositList,        'netsuite/records/customer_refund_deposit_list'
+    autoload :CustomerSubscription,             'netsuite/records/customer_subscription'
+    autoload :CustomerSubscriptionsList,        'netsuite/records/customer_subscriptions_list'
+    autoload :CustomerStatus,                   'netsuite/records/customer_status'
     autoload :CustomerPartner,                  'netsuite/records/customer_partner'
+    autoload :CustomerSalesTeam,                'netsuite/records/customer_sales_team'
+    autoload :CustomerSalesTeamList,            'netsuite/records/customer_sales_team_list'
     autoload :ContactList,                      'netsuite/records/contact_list'
     autoload :Contact,                          'netsuite/records/contact'
+    autoload :ContactAddressbook,               'netsuite/records/contact_addressbook'
+    autoload :ContactAddressbookList,           'netsuite/records/contact_addressbook_list'
     autoload :ContactRole,                      'netsuite/records/contact_role'
     autoload :ContactAccessRoles,               'netsuite/records/contact_access_roles'
     autoload :ContactAccessRolesList,           'netsuite/records/contact_access_roles_list'
@@ -122,6 +146,7 @@ module NetSuite
     autoload :CurrencyRate,                     'netsuite/records/currency_rate'
     autoload :Department,                       'netsuite/records/department'
     autoload :Deposit,                          'netsuite/records/deposit'
+    autoload :DepositApplication,               'netsuite/records/deposit_application'
     autoload :DepositPayment,                   'netsuite/records/deposit_payment'
     autoload :DepositPaymentList,               'netsuite/records/deposit_payment_list'
     autoload :DepositOther,                     'netsuite/records/deposit_other'
@@ -132,16 +157,29 @@ module NetSuite
     autoload :DiscountItem,                     'netsuite/records/discount_item'
     autoload :Duration,                         'netsuite/records/duration'
     autoload :Employee,                         'netsuite/records/employee'
+    autoload :EntityCustomField,                'netsuite/records/entity_custom_field'
     autoload :File,                             'netsuite/records/file'
     autoload :GiftCertificate,                  'netsuite/records/gift_certificate'
     autoload :GiftCertificateItem,              'netsuite/records/gift_certificate_item'
     autoload :GiftCertRedemption,               'netsuite/records/gift_cert_redemption'
     autoload :GiftCertRedemptionList,           'netsuite/records/gift_cert_redemption_list'
     autoload :Folder,                           'netsuite/records/folder'
+    autoload :InboundShipment,                  'netsuite/records/inbound_shipment'
+    autoload :InboundShipmentItem,              'netsuite/records/inbound_shipment_item'
+    autoload :InboundShipmentItemList,          'netsuite/records/inbound_shipment_item_list'
+    autoload :InterCompanyJournalEntry,         'netsuite/records/inter_company_journal_entry'
+    autoload :InterCompanyJournalEntryLine,     'netsuite/records/inter_company_journal_entry_line'
+    autoload :InterCompanyJournalEntryLineList, 'netsuite/records/inter_company_journal_entry_line_list'
+    autoload :InventoryAdjustment,              'netsuite/records/inventory_adjustment'
+    autoload :InventoryAdjustmentInventory,     'netsuite/records/inventory_adjustment_inventory'
+    autoload :InventoryAdjustmentInventoryList, 'netsuite/records/inventory_adjustment_inventory_list'
     autoload :InventoryAssignment,              'netsuite/records/inventory_assignment'
     autoload :InventoryAssignmentList,          'netsuite/records/inventory_assignment_list'
     autoload :InventoryDetail,                  'netsuite/records/inventory_detail'
     autoload :InventoryItem,                    'netsuite/records/inventory_item'
+    autoload :InventoryNumber,                  'netsuite/records/inventory_number'
+    autoload :InventoryNumberLocations,         'netsuite/records/inventory_number_locations'
+    autoload :InventoryNumberLocationsList,     'netsuite/records/inventory_number_locations_list'
     autoload :InventoryTransfer,                'netsuite/records/inventory_transfer'
     autoload :InventoryTransferInventory,       'netsuite/records/inventory_transfer_inventory'
     autoload :InventoryTransferInventoryList,   'netsuite/records/inventory_transfer_inventory_list'
@@ -169,8 +207,11 @@ module NetSuite
     autoload :KitItem,                          'netsuite/records/kit_item'
     autoload :Location,                         'netsuite/records/location'
     autoload :LocationsList,                    'netsuite/records/locations_list'
+    autoload :LotNumberedAssemblyItem,          'netsuite/records/lot_numbered_assembly_item'
+    autoload :LotNumberedInventoryItem,         'netsuite/records/lot_numbered_inventory_item'
     autoload :MatrixOptionList,                 'netsuite/records/matrix_option_list'
     autoload :MemberList,                       'netsuite/records/member_list'
+    autoload :Message,                          'netsuite/records/message'
     autoload :NonInventorySaleItem,             'netsuite/records/non_inventory_sale_item'
     autoload :NonInventoryPurchaseItem,         'netsuite/records/non_inventory_purchase_item'
     autoload :NonInventoryResaleItem,           'netsuite/records/non_inventory_resale_item'
@@ -185,8 +226,14 @@ module NetSuite
     autoload :PaymentMethod,                    'netsuite/records/payment_method'
     autoload :PayrollItem,                      'netsuite/records/payroll_item'
     autoload :PhoneCall,                        'netsuite/records/phone_call'
+    autoload :Price,                            'netsuite/records/price'
+    autoload :PriceLevel,                       'netsuite/records/price_level'
+    autoload :PriceList,                        'netsuite/records/price_list'
+    autoload :Pricing,                          'netsuite/records/pricing'
     autoload :PricingMatrix,                    'netsuite/records/pricing_matrix'
     autoload :PromotionCode,                    'netsuite/records/promotion_code'
+    autoload :PromotionsList,                   'netsuite/records/promotions_list'
+    autoload :Promotions,                       'netsuite/records/promotions'
     autoload :PurchaseOrder,                    'netsuite/records/purchase_order'
     autoload :PurchaseOrderItemList,            'netsuite/records/purchase_order_item_list'
     autoload :PurchaseOrderItem,                'netsuite/records/purchase_order_item'
@@ -197,24 +244,35 @@ module NetSuite
     autoload :RevRecSchedule,                   'netsuite/records/rev_rec_schedule'
     autoload :RoleList,                         'netsuite/records/role_list'
     autoload :ReturnAuthorization,              'netsuite/records/return_authorization'
+    autoload :ReturnAuthorizationItem,          'netsuite/records/return_authorization_item'
+    autoload :ReturnAuthorizationItemList,      'netsuite/records/return_authorization_item_list'
     autoload :SalesOrder,                       'netsuite/records/sales_order'
     autoload :SalesOrderShipGroupList,          'netsuite/records/sales_order_ship_group_list'
     autoload :SalesOrderItem,                   'netsuite/records/sales_order_item'
     autoload :SalesOrderItemList,               'netsuite/records/sales_order_item_list'
+    autoload :SalesRole,                        'netsuite/records/sales_role'
     autoload :SalesTaxItem,                     'netsuite/records/sales_tax_item'
     autoload :ServiceResaleItem,                'netsuite/records/service_resale_item'
     autoload :ServiceSaleItem,                  'netsuite/records/service_sale_item'
+    autoload :SerializedAssemblyItem,          'netsuite/records/serialized_assembly_item'
     autoload :SerializedInventoryItem,          'netsuite/records/serialized_inventory_item'
+    autoload :SerializedInventoryItemNumbers,       'netsuite/records/serialized_inventory_item_numbers'
+    autoload :SerializedInventoryItemNumbersList,   'netsuite/records/serialized_inventory_item_numbers_list'
+    autoload :SerializedInventoryItemLocation,       'netsuite/records/serialized_inventory_item_location'
+    autoload :SerializedInventoryItemLocationsList,       'netsuite/records/serialized_inventory_item_locations_list'
     autoload :ShipAddress,                      'netsuite/records/ship_address'
     autoload :SiteCategory,                     'netsuite/records/site_category'
     autoload :Subsidiary,                       'netsuite/records/subsidiary'
     autoload :SubtotalItem,                     'netsuite/records/subtotal_item'
     autoload :SupportCase,                      'netsuite/records/support_case'
+    autoload :SupportCaseType,                  'netsuite/records/support_case_type'
     autoload :TaxType,                          'netsuite/records/tax_type'
     autoload :TaxGroup,                         'netsuite/records/tax_group'
     autoload :Task,                             'netsuite/records/task'
     autoload :Term,                             'netsuite/records/term'
     autoload :TimeBill,                         'netsuite/records/time_bill'
+    autoload :TransactionBodyCustomField,       'netsuite/records/transaction_body_custom_field'
+    autoload :TransactionColumnCustomField,     'netsuite/records/transaction_column_custom_field'
     autoload :TransactionShipGroup,             'netsuite/records/transaction_ship_group'
     autoload :TransferOrder,                    'netsuite/records/transfer_order'
     autoload :TransferOrderItemList,            'netsuite/records/transfer_order_item_list'
@@ -236,6 +294,9 @@ module NetSuite
     autoload :VendorCreditItemList,             'netsuite/records/vendor_credit_item_list'
     autoload :VendorCreditExpense,              'netsuite/records/vendor_credit_expense'
     autoload :VendorCreditExpenseList,          'netsuite/records/vendor_credit_expense_list'
+    autoload :VendorReturnAuthorization,        'netsuite/records/vendor_return_authorization'
+    autoload :VendorReturnAuthorizationItem,    'netsuite/records/vendor_return_authorization_item'
+    autoload :VendorReturnAuthorizationItemList, 'netsuite/records/vendor_return_authorization_item_list'
     autoload :VendorPayment,                    'netsuite/records/vendor_payment'
     autoload :VendorPaymentApply,               'netsuite/records/vendor_payment_apply'
     autoload :VendorPaymentApplyList,           'netsuite/records/vendor_payment_apply_list'
@@ -251,25 +312,6 @@ module NetSuite
 
   def self.configure(&block)
     NetSuite::Configuration.instance_eval(&block)
-  end
-
-  def self.configure_from_env(&block)
-    NetSuite.configure do
-      reset!
-
-      email         ENV['NETSUITE_EMAIL']     unless ENV['NETSUITE_EMAIL'].nil?
-      password      ENV['NETSUITE_PASSWORD']  unless ENV['NETSUITE_PASSWORD'].nil?
-      account       ENV['NETSUITE_ACCOUNT']   unless ENV['NETSUITE_ACCOUNT'].nil?
-      role          ENV['NETSUITE_ROLE']      unless ENV['NETSUITE_ROLE'].nil?
-      api_version   ENV['NETSUITE_API']       unless ENV['NETSUITE_API'].nil?
-      sandbox       (ENV['NETSUITE_PRODUCTION'].nil? || ENV['NETSUITE_PRODUCTION'] != 'true')
-      wsdl          ENV['NETSUITE_WSDL']      unless ENV['NETSUITE_WSDL'].nil?
-      silent        (!ENV['NETSUITE_SILENT'].nil? && ENV['NETSUITE_SILENT'] == 'true')
-
-      read_timeout  100_000
-    end
-
-    self.configure(&block) if block
   end
 
 end
