@@ -10,7 +10,7 @@ module NetSuite
       actions :get, :get_list, :add, :update, :delete, :search, :upsert
 
       fields :acct_name, :acct_number, :acct_type, :cash_flow_rate, :cur_doc_num, :description, :eliminate, :exchange_rate,
-        :general_rate, :include_children, :inventory, :is_inactive, :opening_balance, :revalue, :tran_date
+        :general_rate, :include_children, :inventory, :is_inactive, :opening_balance, :revalue, :tran_date, :balance
 
       record_refs :billable_expenses_acct, :category1099misc, :currency, :deferral_acct, :department, :klass, :location, :parent
 
@@ -18,6 +18,7 @@ module NetSuite
 
       attr_reader :internal_id
       attr_accessor :external_id
+      attr_accessor :search_joins
 
       def initialize(attributes = {})
         @internal_id = attributes.delete(:internal_id) || attributes.delete(:@internal_id)
