@@ -6,7 +6,7 @@ module NetSuite
       include Support::Records
       include Namespaces::TranCust
 
-      fields :amount, :rate, :quantity
+      fields :amount, :rate, :quantity, :is_taxable, :order_line, :line, :description
       field :custom_field_list,    CustomFieldList
 
       record_refs :item, :klass, :price
