@@ -13,7 +13,9 @@ module NetSuite
 
         @custom_fields_assoc = Hash.new
         custom_fields.each do |custom_field|
-          # not all custom fields have an id; https://github.com/NetSweet/netsuite/issues/182
+          # not all custom fields have an id
+          # https://github.com/NetSweet/netsuite/issues/182
+
           if reference_id = custom_field.send(reference_id_type)
             @custom_fields_assoc[reference_id.to_sym] = custom_field
           end
@@ -25,7 +27,12 @@ module NetSuite
       end
 
       def delete_custom_field(field)
-        custom_fields.delete_if { |c| c.send(reference_id_type).to_sym == field }
+        custom_fields.delete_if do |c|
+          # https://github.com/NetSweet/netsuite/issues/325
+          c.send(reference_id_type) &&
+            c.send(reference_id_type).to_sym == field
+        end
+
         @custom_fields_assoc.delete(field)
       end
 
@@ -91,17 +98,28 @@ module NetSuite
       end
 
       private
+
         def reference_id_type
           @reference_id_type ||= Configuration.api_version >= '2013_2' ? :script_id : :internal_id
         end
 
         def extract_custom_field(custom_field_data)
-          # TODO this seems brittle, but might sufficient, watch out for this if something breaks
-          if (custom_field_data[:"@xsi:type"] || custom_field_data[:type]) == "platformCore:SelectCustomFieldRef"
-            custom_field_data[:value] = CustomRecordRef.new(custom_field_data.delete(:value))
-          end
+          if custom_field_data.kind_of?(CustomField)
+            custom_fields << custom_field_data
+          else
+            attrs = custom_field_data.clone
+            type = (custom_field_data[:"@xsi:type"] || custom_field_data[:type])
 
-          custom_fields << CustomField.new(custom_field_data)
+            if type == "platformCore:SelectCustomFieldRef"
+              attrs[:value] = CustomRecordRef.new(custom_field_data[:value])
+            elsif type == 'platformCore:MultiSelectCustomFieldRef'
+              attrs[:value] = custom_field_data[:value].map do |entry|
+                CustomRecordRef.new(entry)
+              end
+            end
+
+            custom_fields << CustomField.new(attrs)
+          end
         end
 
         def create_custom_field(reference_id, field_value)
