@@ -13,7 +13,7 @@ module NetSuite
 
       fields :account_number, :aging, :alt_email, :alt_name, :alt_phone, :bill_pay,
         :buying_reason, :buying_time_frame, :campaign_category, :click_stream, :comments, :company_name,
-        :consol_aging, :consol_days_overdue, :contrib_pct, :credit_cards_list, :credit_hold_override,
+        :consol_aging, :consol_days_overdue, :contrib_pct, :credit_hold_override,
         :credit_limit, :date_created, :days_overdue, :default_address,
         :download_list, :email, :email_preference, :email_transactions, :end_date, :entity_id,
         :estimated_budget, :fax, :fax_transactions, :first_name, :first_visit, :give_access, :global_subscription_status,
@@ -22,24 +22,25 @@ module NetSuite
         :opening_balance, :opening_balance_account, :opening_balance_date,
         :password, :password2, :phone, :phonetic_name, :pref_cc_processor, :print_on_check_as,
         :print_transactions, :referrer, :reminder_days, :representing_subsidiary, :require_pwd_change, :resale_number,
-        :sales_group, :sales_readiness, :sales_team_list, :salutation, :send_email, :ship_complete, :shipping_item,
-        :stage, :start_date, :sync_partner_teams, :tax_exempt, :tax_item, :taxable,
+        :sales_group, :sales_readiness, :salutation, :send_email, :ship_complete,
+        :stage, :start_date, :sync_partner_teams, :tax_exempt, :taxable,
         :third_party_acct, :third_party_country, :third_party_zipcode, :title, :url,
         :vat_reg_number, :visits, :web_lead
 
       field :addressbook_list,  CustomerAddressbookList
+      field :credit_cards_list, CustomerCreditCardsList
       field :custom_field_list, CustomFieldList
       field :contact_roles_list, ContactAccessRolesList
       field :currency_list, CustomerCurrencyList
       field :partners_list, CustomerPartnersList
-      
-      # TODO subscriptions_list
+      field :subscriptions_list, CustomerSubscriptionsList
+      field :sales_team_list, CustomerSalesTeamList
 
       read_only_fields :balance, :consol_balance, :deposit_balance, :consol_deposit_balance, :overdue_balance,
         :consol_overdue_balance, :unbilled_orders, :consol_unbilled_orders
 
       record_refs :access_role, :custom_form, :currency, :entity_status, :partner, :category, :lead_source,
-        :price_level,:sales_rep, :subsidiary, :terms, :parent, :territory
+        :price_level, :sales_rep, :subsidiary, :terms, :parent, :territory, :tax_item, :shipping_item, :receivables_account
 
       attr_reader   :internal_id
       attr_accessor :external_id
