@@ -9,7 +9,7 @@ module NetSuite
 
       actions :get, :add, :delete, :search, :update, :upsert, :upsert_list
 
-      fields :created_date, :last_modified_date, :tran_date, :tran_id, :memo
+      fields :klass, :created_date, :last_modified_date, :tran_date, :tran_id, :memo
 
       field :inventory_list, InventoryTransferInventoryList
 
