@@ -7,7 +7,7 @@ module NetSuite
       include Support::Actions
       include Namespaces::TranPurch
 
-      actions :get, :get_list, :add, :initialize, :delete, :update, :upsert, :search
+      actions :get, :get_deleted, :get_list, :add, :initialize, :delete, :update, :upsert, :search
 
       fields :created_date, :currency_name, :exchange_rate, :landed_cost_per_line,
              :last_modified_date, :memo, :tran_date, :tran_id
