@@ -15,10 +15,14 @@ module NetSuite
       #
       #   <listAcct:matrixOptionList>
       #     <listAcct:matrixOption internalId="45" scriptId="custitem13">
-      #       <platformCore:value internalId="4" typeId="28"/>
+      #       <platformCore:value internalId="4" typeId="28">
+      #         <platformCore:name>foo</platformCore:name>
+      #       </platformCore:value>
       #     </listAcct:matrixOption>
       #     <listAcct:matrixOption internalId="46" scriptId="custitem14">
-      #       <platformCore:value internalId="1" typeId="29"/>
+      #       <platformCore:value internalId="1" typeId="29">
+      #         <platformCore:name>bar</platformCore:name>
+      #       </platformCore:value>
       #     </listAcct:matrixOption>
       #   </listAcct:matrixOptionList>
       #
@@ -27,13 +31,17 @@ module NetSuite
         when Hash
           options << OpenStruct.new(
             type_id: attributes[:matrix_option][:value][:'@type_id'],
-            value_id: attributes[:matrix_option][:value][:'@internal_id']
+            value_id: attributes[:matrix_option][:value][:'@internal_id'],
+            script_id: attributes[:matrix_option][:@script_id],
+            name: attributes[:matrix_option][:value][:name]
           )
         when Array
           attributes[:matrix_option].each do |option|
             options << OpenStruct.new(
               type_id: option[:value][:'@type_id'],
-              value_id: option[:value][:'@internal_id']
+              value_id: option[:value][:'@internal_id'],
+              script_id: option[:@script_id],
+              name: option[:value][:name]
             )
           end
         end
