@@ -6,11 +6,18 @@ module NetSuite
       include Support::Records
       include Namespaces::TranSales
 
-      fields :amount, :bin_numbers, :cost_estimate, :cost_estimate_type, :defer_rev_rec, :description, :gift_cert_from,
-        :gift_cert_message, :gift_cert_number, :gift_cert_recipient_email, :gift_cert_recipient_name, :gross_amt, :is_taxable,
-        :line, :order_line, :po_currency, :quantity, :rate, :rev_rec_end_date, :rev_rec_start_date, :rev_rec_term_in_months,
-        :serial_numbers, :shipping_cost, :tax1_amt, :tax_rate1, :tax_rate2, :vsoe_allocation, :vsoe_amount, :vsoe_deferral,
-        :vsoe_delivered, :vsoe_permit_discount, :vsoe_price, :is_closed, :quantity_commited, :quantity_fulfilled
+      fields :amount, :bin_numbers, :cost_estimate,
+        :cost_estimate_type, :defer_rev_rec, :description,
+        :expand_item_group, :gift_cert_from, :gift_cert_message,
+        :gift_cert_number, :gift_cert_recipient_email,
+        :gift_cert_recipient_name, :gross_amt, :is_closed,
+        :is_taxable, :line, :order_line, :po_currency, :quantity,
+        :quantity_back_ordered, :quantity_billed, :quantity_committed,
+        :quantity_fulfilled, :rate, :rev_rec_end_date,
+        :rev_rec_start_date, :rev_rec_term_in_months, :serial_numbers,
+        :shipping_cost, :tax1_amt, :tax_rate1, :tax_rate2,
+        :vsoe_allocation, :vsoe_amount, :vsoe_deferral,
+        :vsoe_delivered, :vsoe_permit_discount, :vsoe_price
 
       field :custom_field_list, CustomFieldList
 
