@@ -1,3 +1,3 @@
-module Netsuite
-  VERSION = '0.7.5.flexport'
+module NetSuite
+  VERSION = '0.8.5.flexport'
 end
