@@ -8,7 +8,7 @@ describe NetSuite::Records::Customer do
       :account_number, :aging, :alt_email, :alt_name, :alt_phone, :balance, :bill_pay,
       :buying_reason, :buying_time_frame, :campaign_category, :click_stream, :comments, :company_name,
       :consol_aging, :consol_balance, :consol_days_overdue, :consol_deposit_balance, :consol_overdue_balance,
-      :consol_unbilled_orders, :contrib_pct, :credit_cards_list, :credit_hold_override, :credit_limit,
+      :consol_unbilled_orders, :contrib_pct, :credit_hold_override, :credit_limit,
       :date_created, :days_overdue, :default_address,
       :deposit_balance, :download_list, :email, :email_preference, :email_transactions, :end_date, :entity_id,
       :estimated_budget, :fax, :fax_transactions, :first_name, :first_visit, :give_access, :global_subscription_status,
@@ -17,8 +17,8 @@ describe NetSuite::Records::Customer do
       :opening_balance, :opening_balance_account, :opening_balance_date, :overdue_balance,
       :password, :password2, :phone, :phonetic_name, :pref_cc_processor,:print_on_check_as,
       :print_transactions, :referrer, :reminder_days, :representing_subsidiary, :require_pwd_change, :resale_number,
-      :sales_group, :sales_readiness, :sales_team_list, :salutation, :send_email, :ship_complete, :shipping_item,
-      :stage, :start_date, :sync_partner_teams, :tax_exempt, :tax_item, :taxable,
+      :sales_group, :sales_readiness, :salutation, :send_email, :ship_complete,
+      :stage, :start_date, :sync_partner_teams, :tax_exempt, :taxable,
       :third_party_acct, :third_party_country, :third_party_zipcode, :title, :unbilled_orders, :url,
       :vat_reg_number, :visits, :web_lead
     ].each do |field|
@@ -28,7 +28,7 @@ describe NetSuite::Records::Customer do
 
   it 'has the right record_refs' do
     [
-      :access_role, :currency, :custom_form, :entity_status, :partner, :sales_rep, :terms, :parent, :territory
+      :access_role, :currency, :custom_form, :entity_status, :partner, :sales_rep, :terms, :parent, :territory, :shipping_item, :tax_item
     ].each do |record_ref|
       expect(customer).to have_record_ref(record_ref)
     end
@@ -65,6 +65,27 @@ describe NetSuite::Records::Customer do
     end
   end
 
+  describe '#credit_cards_list' do
+    it 'can be set from attributes' do
+      customer.credit_cards_list = {
+        :credit_cards => {
+          :internal_id       => '1234567',
+          :cc_default        => true,
+          :cc_expire_date    => '2099-12-01T00:00:00.000-08:00'
+        }
+      }
+
+      expect(customer.credit_cards_list).to be_kind_of(NetSuite::Records::CustomerCreditCardsList)
+      expect(customer.credit_cards_list.credit_cards.length).to eql(1)
+    end
+
+    it 'can be set from a CustomerCreditCardsList object' do
+      customer_credit_cards_list = NetSuite::Records::CustomerCreditCardsList.new
+      customer.credit_cards_list = customer_credit_cards_list
+      expect(customer.credit_cards_list).to eql(customer_credit_cards_list)
+    end
+  end
+
   describe '#custom_field_list' do
     it 'can be set from attributes' do
       attributes = {
@@ -84,6 +105,27 @@ describe NetSuite::Records::Customer do
       expect(customer.custom_field_list).to eql(custom_field_list)
     end
   end
+
+  describe '#subscriptions_list' do
+    it 'can be set from attributes' do
+      customer.subscriptions_list = {
+        subscriptions: [
+          {
+            :subscribed         => true
+          }
+        ]
+      }
+      expect(customer.subscriptions_list).to be_kind_of(NetSuite::Records::CustomerSubscriptionsList)
+      expect(customer.subscriptions_list.subscriptions.length).to eql(1)
+    end
+
+    it 'can be set from a CustomerSubscriptionsList object' do
+      customer_subscriptions_list = NetSuite::Records::CustomerSubscriptionsList.new
+      customer.subscriptions_list = customer_subscriptions_list
+      expect(customer.subscriptions_list).to eql(customer_subscriptions_list)
+    end
+  end
+
 
   describe '.get' do
     context 'when the response is successful' do
