@@ -1,5 +1,10 @@
 $LOAD_PATH.unshift(File.expand_path(File.join(File.dirname(__FILE__), '..') ))
 
+require "rubygems"
+require "bundler/setup"
+
+Bundler.require
+
 # https://circleci.com/docs/code-coverage
 if ENV['CIRCLE_ARTIFACTS']
   require 'simplecov'
@@ -7,10 +12,6 @@ if ENV['CIRCLE_ARTIFACTS']
   SimpleCov.coverage_dir(dir)
   SimpleCov.start
 end
-
-require 'rspec'
-require 'netsuite'
-require 'pry'
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
